wip working on sql query with count result

+++ Commons/src/main/java/eu/neclab/ngsildbroker/commons/querybase/BaseQueryService.java
@@ -196,7 +196,7 @@ public abstract class BaseQueryService implements EntryQueryService {
 							HttpEntity<String> entity;
 							String resultBody;
 							ResponseEntity<String> response;
-							int count = 0;
+							long count = 0;
 							if (postQuery) {
 								entity = new HttpEntity<String>(rawQueryString, additionalHeaders);
 								response = restTemplate.exchange(endpoint + "/ngsi-ld/v1/entityOperations/query",
@@ -210,8 +210,8 @@ public abstract class BaseQueryService implements EntryQueryService {
 								resultBody = response.getBody();
 							}
 							if (response.getHeaders().containsKey(NGSIConstants.COUNT_HEADER_RESULT)) {
-								count = Integer
-										.parseInt(response.getHeaders().get(NGSIConstants.COUNT_HEADER_RESULT).get(0));
+								count = Long
+										.parseLong(response.getHeaders().get(NGSIConstants.COUNT_HEADER_RESULT).get(0));
 							}
 							logger.debug("http call result :: ::" + resultBody);
 
@@ -311,7 +311,7 @@ public abstract class BaseQueryService implements EntryQueryService {
 		ExecutorService executorService = Executors.newFixedThreadPool(2);
 		List<Future<RemoteQueryResult>> futures = executorService.invokeAll(callablesCollection);
 		RemoteQueryResult queryResult = new RemoteQueryResult(null, ErrorType.None, -1, true);
-		int count = 0;
+		long count = 0;
 		for (Future<RemoteQueryResult> future : futures) {
 			logger.trace("future.isDone = " + future.isDone());
 			RemoteQueryResult tempResult = future.get();
